P5c: split joystick_controller into state machine + joystick_panels (facade preserved)
The 4,700-line module mixed the safety-critical tracking state machine
with ~3,000 lines of panel rendering. Now:

- joystick_controller.py (1,800 lines): TrackingMode, JoystickModeState
  (tracking control, park, hotspot, focus, joystick input), layouts,
  axis mapping -- the half the control thread and the Rust adapter
  depend on.
- joystick_panels.py (3,000 lines): every render_* panel, feed
  processing/caching, navball, strip charts, plate-solve panel, and
  their mouse handlers.

joystick_controller re-imports every name joystick_panels defines
(including private ones), so it remains a complete facade: all existing
imports (main.py's 20-name import, alignment_ui's private-helper
imports, make_live_screenshots) work unchanged. Verified by pyflakes
(no undefined names), the full 294-test suite, the A/B loop harness,
and a headless runtime smoke that executes all 14 major panel
renderers against a real DisplaySetup.
